Moved image scaling things into subdir
