+++ src/components/Global/Scaling/CompositeImage.vue
