+++ src/components/Global/Scaling/HistogramSlider.vue
