+++ src/components/Global/Scaling/ImageScaler.vue
