+++ src/components/Global/Scaling/RawScaledImage.vue
